refactoring - removing a bunch of unused code
Earlier in the development process, it seemed like a good idea to try
different strategies for earlier guesses than for later guesses.  Now
there is enough test data to show that only one of the strategies I
developed (middle score) solves all the test cases, and it must be used
for every guess.  As a result, I'm removing a bunch of code that appears
useless now.  I'm still leaving the guess counter in place, even though
it is not currently being used in recommend_guess.  It may be needed at
a later data to improve the algorithm if we get a test case that can't
be solved by the existing algorithm.