ran against some large datasets
I did many runs of 100k iterations, and I now believe this is a better
algorithm than the previous version, as there are significantly fewer
failures.

The average number of guesses per test is slightly higher with this
algorithm.  But that is OK, as we want to optimize for an algorithm that
will solve more cases, not one that will solve a given test in less
guesses.

Here are the failure rates for 100k data points of randomly generated
words and answers of various lengths:

  5 letter words - 5%
  6 letter words - 3%
  7 letter words - 2%