next generation recommendation algorithm
This commit throws away all the previous work to score words based on
similarity.

The new algorithm runs a simulation in which every word is considered as
the potential answer.  It plays out the simulation for each answer,
trying to guess using every other word in every valid permutation.  The
key here is to identify all the paths that lead to failure, i.e. where
the simulated answer could not successfully be guessed within four
tries.  That information is then used to pick a next guess word that has
the least potential chance of failure.  The chance for success is
assured if a guess has no failure paths associated with it, i.e. it led
to successfully guessing every other remaining word in the simulation.
If the chance of success is not assured, then some feedback is given
about the number of failure paths that remain.

This change has both reduced the number of failures and the number of
guesses required when run over a large random dataset.